Refactor the server services into a folder structure
