+++ srcs/frontend/server-services/index.js
@@ -1,0 +1,3 @@
+const socketListeners = require("./socket-listeners/index.js")
+
+module.exports = { socketListeners };
